refactor: reorganize utility imports under lib directory
Moved utility modules from the utils directory to the lib directory for better organization and structure. Updated all related imports across the codebase to reflect the new file paths. Minor change to default config to adjust exportPathPrefixesToRemove.
