+++ scripts/update-exports/src/lib/file-utils.ts
