+++ scripts/update-exports/src/lib/logger.ts
