+++ scripts/update-exports/src/lib/path-utils.ts
